requirements.txt
aioredis==1.3.1
amqp==2.6.1
APScheduler==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
daphne==2.5.0
Django==2.2.27
django-apscheduler==0.5.1
django-cors-headers==3.4.0
django-easy-audit==1.3.0
django-filter==2.4.0
django-redis==4.12.1
django-simpleui==2020.9
django-user-agents==0.4.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-writable-nested==0.6.1
drf-yasg==1.17.1
eventlet==0.31.0
greenlet==0.4.16
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
importlib-metadata==1.7.0
incremental
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
packaging==20.4
Pillow==9.0.1
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==22.2.0
txaio==20.4.1
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.5
user-agents==2.2.0
vine==1.3.0
zipp==3.1.0
zope.interface==5.1.0
